+++ src/controlers/user.js
@@ -1,4 +1,6 @@
 const { PrismaClient } = require('@prisma/client');
+const { userSalt } = require('../config/middleware');
+
 const prisma = new PrismaClient();
 
 const include = {
@@ -43,19 +45,25 @@ class UserController {
             .catch(err => res.status(500).json({ error: true, message: err }));
     }
 
-    create(req, res) {
+    async create(req, res) {
         const params = req.body;
 
         if(!params) {
             return res.status(400).json({ error: true, message: "Please give an request body" });
         }
 
-        if(!(params.mail)) {
+        if(!params.mail) {
             return res.status(400).json({ error: true, message: 'Please provide an email' });
         }
 
         if(params.id !== undefined) {
             delete params.id;
+        }
+
+        if(params.type <= 2) {
+            const userHash = await userSalt(params.password);
+            params.password = userHash.hash;
+            Object.defineProperty(params, 'salt', { value: userHash.salt, enumerable: true, writable: true });
         }
 
         prisma.user.create({
@@ -78,10 +86,10 @@ class UserController {
     }
 
     async delete(req, res) {
-        // const user = req.user;
-        // if(user.type > 2 || user.user_id != req.params.id) {
-        //     return res.status(403).json({ error: true, message: "Forbiden" });
-        // }
+        const user = req.user;
+        if(user.type > 2 || user.user_id != req.params.id) {
+            return res.status(403).json({ error: true, message: "Forbiden" });
+        }
 
         prisma.user.delete({
             where: { id: req.params.id }
@@ -91,10 +99,10 @@ class UserController {
     }
 
     update(req, res) {
-        // const user = req.user;
-        // if(user.type > 2 || user.user_id != req.params.id) {
-        //     return res.status(403).json({ error: true, message: "Forbiden" });
-        // }
+        const user = req.user;
+        if(user.type > 2 || user.user_id != req.params.id) {
+            return res.status(403).json({ error: true, message: "Forbiden" });
+        }
 
         const id = req.params.id;
         const params = req.body;
